fix: Rename package to `conflicts`
Motivation: Initially we were thinking about this package being general handler for the offlien state including object mapping. However at this point we know that all the code here will be strictly related to conflict resolution. With possibility to reuse some of the types here in client it makes sense to rename it back to apollo-voyager-conflicts